Fixed dynamic resizing. Fixed messagebox.
Fixed to dynamically resize the application.

Added the rowconfigure() and columnconfigure() Grid methods.
Added sticky parameter on widget’s grid() method.



The rowconfigure() and columnconfigure() methods configure the column indexes of a grid.
The weight parameter determines how wide the column will occupy relative to other columns. The weight is relative.
If two columns have the same weight, they will expand at the same rate.

Use sticky to adapt the size of the widget dynamically too.

Sticky by default is centered. You can specify which sides this widget will stick to the cell boundary using the compass directions, NSEW. Example: ticky="EW" 
 will stick to the right and left of the cell only.

If it is a draw, messagebox is not called. Moved the nested condition. Fixed